fix: resolve all 17 reviewer issues - bugs, vulnerabilities, code smells, duplications

+++ src/main/java/com/multigenesystask/entity/Order.java
@@ -65,7 +65,7 @@ public class Order {
 	private PaymentDetails paymentDetails;
 	
 	
-	private double totalPrice;
+	private Integer totalPrice;
 	
 	
 	private Integer totalDiscountedPrice;
